+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Updated
+
+- Updated to show 4 decimal places.
+
 ## [0.1.100] - 2025-05-26
 
 ### Fixed